users controller up and tested

+++ app/helpers/sessions_helper.rb
@@ -14,7 +14,7 @@ module SessionsHelper
   end
 
   def logout
-    session.delete(:user_id)
+    session.delete(:user_id) if session[:user_id]
     @current_user = nil
   end  
 end
